refactor: refactor client options and fix tests
